feat(): edit working and no lint erros but no local update
